Move json stuff to json package
