+++ java/src/main/java/io/vertx/ext/eventbus/client/json/GsonCodec.java
@@ -1,4 +1,4 @@
-package io.vertx.ext.eventbus.client;
+package io.vertx.ext.eventbus.client.json;
 
 import com.google.gson.Gson;
