+++ java/src/main/java/io/vertx/ext/eventbus/client/json/JacksonCodec.java
@@ -1,4 +1,4 @@
-package io.vertx.ext.eventbus.client;
+package io.vertx.ext.eventbus.client.json;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
