+++ java/src/main/java/io/vertx/ext/eventbus/client/json/JsonCodec.java
@@ -1,4 +1,4 @@
-package io.vertx.ext.eventbus.client;
+package io.vertx.ext.eventbus.client.json;
 
 /**
  * @author <a href="mailto:julien@julienviet.com">Julien Viet</a>
