+++ java/src/test/java/io/vertx/ext/eventbus/client/JacksonCodecTest.java
@@ -1,5 +1,7 @@
 package io.vertx.ext.eventbus.client;
 
+import io.vertx.ext.eventbus.client.json.JacksonCodec;
+
 /**
  * @author <a href="mailto:julien@julienviet.com">Julien Viet</a>
  */
